+++ examples/05_selection_motion_group_activation.py
@@ -28,8 +28,7 @@ async def move_robot(motion_group: MotionGroup, tcp: str):
         ptp(home_pose),
     ]
 
-    async for _ in motion_group.plan_and_execute(actions, tcp=tcp):
-        pass
+    await motion_group.plan_and_execute(actions, tcp=tcp)
 
 
 async def main():