+++ nova_rerun_bridge/helper_scripts/unpack_models.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
 for f in draco/*.glb; do
-    gltf-transform copy "$f" "./$(basename "$f")"  
+    gltf-transform copy "$f" "./$(basename "$f")"
 done
 
 for f in *.glb; do
-    gltf-transform unlit "$f" "$f"  
+    gltf-transform unlit "$f" "$f"
 done